src/testbench/requirements.txt
numpy==1.26.2
Pillow==10.1.0
